Standardized Transform2D (lercher-fixme)
Introducing:
-) Point2D
-) Size2D
and Transfom2D (both)
... to unify our systems.

+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/board/BoardJsonClasses.kt
@@ -19,12 +19,12 @@ data class TerritoryData(
 
 @Serializable
 data class Position(
-    val x: Int,
-    val y: Int
+    val x: Float,
+    val y: Float
 )
 
 @Serializable
 data class Size(
-    val x: Int,
-    val y: Int
+    val x: Float,
+    val y: Float
 )
